refactor index module
