+++ search-engine/.devcontainer/devcontainer.json
@@ -2,19 +2,18 @@
   "build": {
     "dockerfile": "../Dockerfile"
   },
-  "forwardPorts": [
-    5000
-  ],
   "customizations": {
     "vscode": {
       "extensions": [
         "ms-python.python",
+        "ms-python.pylint",
+        "ms-python.black-formatter",
         "ms-toolsai.jupyter",
         "ms-toolsai.jupyter-renderers"
       ]
     }
   },
   "mounts": [
-    "source=${localEnv:HOME}/data,target=/mnt/data,type=bind,consistency=cached"
+    "source=${localEnv:HOME}/biosearchdev,target=/mnt,type=bind,consistency=cached"
   ]
 }
