+++ search-engine/.gitignore
@@ -1,0 +1,1 @@
+.pytest_cache
