+++ search-engine/.pylintrc
@@ -1,0 +1,3 @@
+[MASTER]
+init-hook="from pylint.config import find_pylintrc;
+import os, sys; sys.path.append(os.path.dirname(find_pylintrc()))"
