+++ search-engine/.vscode/settings.json
@@ -1,3 +1,6 @@
 {
-  "python.formatting.provider": "black"
+  "python.formatting.provider": "none",
+  "[python]": {
+    "editor.defaultFormatter": "ms-python.black-formatter"
+  }
 }
