+++ search-engine/Dockerfile
@@ -1,4 +1,4 @@
-FROM coady/pylucene:8.11
+FROM coady/pylucene:8
 
 RUN apt update && apt install -y tmux && apt clean && apt autoclean & rm -rf /var/lib/apt/lists/*
 COPY requirements.txt ./requirements.txt
